requirements.txt
streamlit==1.34.0
deepface==0.0.79
opencv-python==4.9.0.80
pillow
numpy
pandas
tensorflow==2.12.0
